requirements.txt
defaults
torch>=1.10.1
torchvision
ninja
diffusers
pudb==2019.2
albumentations==1.3.0
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==1.4.2
omegaconf==2.1.1
torchmetrics==0.6
test-tube>=0.7.5
streamlit>=0.73.1
einops==0.3.0
omegaconf==2.1.1
transformers==4.19.2
webdataset==0.2.5
gradio==3.11
kornia==0.6
open-clip-torch==2.7.0
invisible-watermark>=0.1.5
streamlit-drawable-canvas==0.8.0
numpy==1.23.1
Pillow==9.3.0
torchtext==0.11.1
accelerate
opencv-python==4.5.5.64

